Issue/184 - Move updated date strategy to has more strategies namespace
